requirements.txt
PyQt5>=5.15.0
pynput>=1.7.3
pyautogui>=0.9.53
opencv-python>=4.5.5.62
pytesseract>=0.3.10
Pillow>=8.4.0
numpy>=1.21.0